int->float타입변경

+++ backend/database/migrations/2022_02_16_074845_create_posts_table.php
@@ -20,10 +20,10 @@ class CreatePostsTable extends Migration
                 ->onDelete('cascade');  //같이 삭제 요청
             $table->string('kind');
             $table->string('time');
-            $table->integer('calorie');
+            $table->float('calorie');
             $table->string('average_speed');
             $table->string('altitude');
-            $table->integer('distance');
+            $table->float('distance');
             $table->string('img')->nullable();
             $table->string('content')->nullable();
             $table->string('range');
